add the DataSink trait for writing messages
Add the trait, and implement it for the CBOR format. This requires a
little bit of renaming and shuffling code around.

The GroupHeader trait now requires the from_msg fn, which computes the
correct header value for a Versioned type.

While DataSink is not specific to message groups, it can be implemented
for the same types as DataSource, so it lives in the group module for
now.