add headers that contain the message length
Add cbor::BasicHeader, which contains a 32-bit length field to
communicate how long the message is. This is mostly because serde_cbor
is dumb and doesn't have a mode that can tolerate trailing data, but
it's also a perfectly valid mode for other reasons.

Rename cbor::FixedHeader to TinyHeader.

The addition of the length field set off a bunch of other changes:
- DataSource::read_message accepts a header, which can be useful in some
  cases.
- UpgradeLatest::upgrade_latest accepts a header rather than just a
  version.
- GroupHeader::for_msg is gone. It's impossible to create a header for
  an arbitrary message before it's serialized, because headers that
  contain the message length wouldn't be possible.
- DataSink was simplified; writing a header and message separately
  aren't that useful so it's now just one function write_message.